Refactoring polling to its own package
Just doing some organization.  This will likely change soon when I
figure out how this project should be organized.
